refactor(ai): fix field naming and add XML docs across BehaviourTree nodes (#21)
