+++ UNITY/Assets/Scripts/AI/BehaviourTree/IBehaviourNode.cs
@@ -1,9 +1,9 @@
-using WarOfTanks.Enums;
-
-namespace WarOfTanks.Interfaces
+namespace WarOfTanks.AI.BehaviourTree
 {
+    /// <summary>Defines one executable node in a behaviour tree.</summary>
     public interface IBehaviourNode
     {
+        /// <summary>Runs this node once and returns its current execution status.</summary>
         NodeStatus Tick();
     }
 }
