+++ UNITY/Assets/Scripts/AI/BehaviourTree/Inverter.cs
@@ -1,17 +1,17 @@
-using WarOfTanks.Enums;
-using WarOfTanks.Interfaces;
-
 namespace WarOfTanks.AI.BehaviourTree
 {
+    /// <summary>Decorator that flips success and failure results from its child.</summary>
     public class Inverter : DecoratorNode
     {
+        /// <summary>Creates an inverter around one child node.</summary>
         public Inverter(IBehaviourNode child) : base(child)
         {
         }
 
+        /// <summary>Ticks the child, flips success/failure, and preserves running.</summary>
         public override NodeStatus Tick()
         {
-            var status = child.Tick();
+            var status = _child.Tick();
 
             if (status == NodeStatus.Success)
             {
