+++ UNITY/Assets/Scripts/AI/BehaviourTree/NodeStatus.cs
@@ -1,5 +1,6 @@
-namespace WarOfTanks.Enums
+namespace WarOfTanks.AI.BehaviourTree
 {
+    /// <summary>Represents the result of ticking a behaviour tree node.</summary>
     public enum NodeStatus
     {
         Success,
