+++ UNITY/Assets/Scripts/AI/BehaviourTree/Sequence.cs
@@ -1,18 +1,19 @@
 using System.Collections.Generic;
-using WarOfTanks.Enums;
-using WarOfTanks.Interfaces;
 
 namespace WarOfTanks.AI.BehaviourTree
 {
+    /// <summary>Composite node that succeeds only when all children succeed.</summary>
     public class Sequence : CompositeNode
     {
+        /// <summary>Creates a sequence with the initial child list.</summary>
         public Sequence(List<IBehaviourNode> children) : base(children)
         {
         }
 
+        /// <summary>Ticks children until one fails or runs; succeeds when all children succeed.</summary>
         public override NodeStatus Tick()
         {
-            foreach (var child in children)
+            foreach (var child in _children)
             {
                 var status = child.Tick();
