+++ UNITY/Assets/Tests/BehaviourTreeTests.cs
@@ -1,8 +1,6 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 using WarOfTanks.AI.BehaviourTree;
-using WarOfTanks.Enums;
-using WarOfTanks.Interfaces;
 
 public class BehaviourTreeTests
 {
